+++ client/package.json
@@ -26,6 +26,7 @@
     "react": "^17.0.1",
     "react-day-picker": "^7.4.10",
     "react-dom": "^17.0.1",
+    "react-material-ui-carousel": "^2.2.6",
     "react-multi-date-picker": "^2.9.8",
     "react-router-dom": "^5.2.0",
     "react-scripts": "^4.0.0",
