+++ client/src/components/NavBar/NavBar.tsx
@@ -103,7 +103,7 @@ const NavBar = (): JSX.Element => {
                 </Link>
                 <Badge
                   color="secondary"
-                  badgeContent={unreadNotifications.length}
+                  badgeContent={0}
                   max={99}
                   anchorOrigin={{ vertical: 'top', horizontal: 'right' }}
                 >
@@ -134,13 +134,13 @@ const NavBar = (): JSX.Element => {
                   }}
                 >
                   <List>
-                    {unreadNotifications.length > 0 ? (
+                    {/* {unreadNotifications.length > 0 ? (
                       unreadNotifications.map((notification) => (
                         <Notification key={notification._id} title={notification.title} content={notification.content} />
                       ))
                     ) : (
                       <MenuItem>You have no unread notifications</MenuItem>
-                    )}
+                    )} */}
                   </List>
                 </Popover>
                 <Badge color="primary" variant="dot">
