+++ server/app.js
@@ -17,6 +17,7 @@ const notificationsRouter = require("./routes/notifications")
 const { json, urlencoded } = express;
 const cors = require('cors')
 
+
 connectDB();
 const app = express();
 const server = http.createServer(app);
